+++ tullamoreqa-core-api/src/main/java/com/gkenna/tullamoreqa/core/api/controllers/QuestionController.java
@@ -7,6 +7,8 @@ package com.gkenna.tullamoreqa.core.api.controllers;
 import com.gkenna.tullamoreqa.domain.Question;
 import org.springframework.http.ResponseEntity;
 
+import java.math.BigInteger;
+
 /**
  * API Controller for the {@link Question} Entity. This API will allow
  * external parties, i.e. UI or CLI, to Get/Add/Update/Delete Questions.
@@ -37,7 +39,7 @@ public interface QuestionController {
      * @param questionId The ID of the {@link Question} to retrieve.
      * @return The Response of this Request.
      */
-    ResponseEntity<Question> getQuestion(final Long questionId);
+    ResponseEntity<Question> getQuestion(final BigInteger questionId);
 
     /**
      * HTTP PUT Method
@@ -49,7 +51,7 @@ public interface QuestionController {
      *                   new values for questionId to update to.
      * @return The Response of this Request.
      */
-    ResponseEntity<?> updateQuestion(final Long questionId,
+    ResponseEntity<?> updateQuestion(final BigInteger questionId,
                                      final Question input);
 
     /**
@@ -60,5 +62,5 @@ public interface QuestionController {
      * @param questionId The ID of the {@link Question} to delete.
      * @return The Response of this Request.
      */
-    ResponseEntity<?> deleteQuestion(final Long questionId);
+    ResponseEntity<?> deleteQuestion(final BigInteger questionId);
 }
