+++ tullamoreqa-core-api/src/main/java/com/gkenna/tullamoreqa/core/api/repositories/AnswerRepository.java
@@ -11,6 +11,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 /**
  * Repository for containing {@link Answer}s. Will be called by the responsible
  * Service, in this case it will be
@@ -20,7 +22,7 @@ import org.springframework.stereotype.Repository;
  * @since 0.0.0
  */
 @Repository
-public interface AnswerRepository extends JpaRepository<Answer, Long> {
+public interface AnswerRepository extends JpaRepository<Answer, BigInteger> {
     /**
      * Return a list of all {@link Answer}s associated with a particular
      * {@link com.gkenna.tullamoreqa.domain.Question}.
@@ -30,7 +32,7 @@ public interface AnswerRepository extends JpaRepository<Answer, Long> {
      * @param pageable   Potentially add Pagination.
      * @return List of {@link Answer}s in Page form.
      */
-    Page<Answer> findByQuestionId(Long questionId, Pageable pageable);
+    Page<Answer> findByQuestionId(BigInteger questionId, Pageable pageable);
 
     /**
      * Return a list of all {@link Answer}s associated with a particular
