+++ tullamoreqa-core-api/src/main/java/com/gkenna/tullamoreqa/core/api/repositories/CommentRepository.java
@@ -8,6 +8,8 @@ import com.gkenna.tullamoreqa.domain.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 /**
  * Repository for containing {@link Comment}s. Will be called by the responsible
  * Service, in this case it will be
@@ -17,5 +19,5 @@ import org.springframework.stereotype.Repository;
  * @since 0.0.0
  */
 @Repository
-public interface CommentRepository extends JpaRepository<Comment, Long> {
+public interface CommentRepository extends JpaRepository<Comment, BigInteger> {
 }
