+++ tullamoreqa-core-api/src/main/java/com/gkenna/tullamoreqa/core/api/services/CommentService.java
@@ -9,6 +9,8 @@ import com.gkenna.tullamoreqa.domain.Comment;
 import com.gkenna.tullamoreqa.domain.User;
 import org.springframework.stereotype.Service;
 
+import java.math.BigInteger;
+
 /**
  * API for interacting with {@link com.gkenna.tullamoreqa.domain.Comment}.
  * This API communicates directly with the
@@ -44,7 +46,7 @@ public interface CommentService {
      * @throws CommentNotFoundException Thrown if the {@link Comment}
      *                                  isn't found.
      */
-    Comment deleteComment(final Long commentId) throws CommentNotFoundException;
+    Comment deleteComment(final BigInteger commentId) throws CommentNotFoundException;
 
     /**
      * Return if a {@link Comment} exists in the DB or not.
@@ -62,7 +64,7 @@ public interface CommentService {
      * @return True if the {@link Comment} exists,
      * false if otherwise.
      */
-    boolean doesCommentExist(final Long commentId);
+    boolean doesCommentExist(final BigInteger commentId);
 
     /**
      * Return a {@link Comment} based on its ID.
@@ -70,7 +72,7 @@ public interface CommentService {
      * @param commentId The ID of the {@link Comment} to return.
      * @return A {@link Comment} with the ID passed.
      */
-    Comment getComment(final Long commentId);
+    Comment getComment(final BigInteger commentId);
 
     /**
      * Return all {@link Comment}s in the DB.
@@ -99,6 +101,6 @@ public interface CommentService {
      * @throws CommentNotFoundException Thrown when the {@link Comment} cannot
      *                                  be found.
      */
-    Comment updateComment(final Long commentId, final Comment input)
+    Comment updateComment(final BigInteger commentId, final Comment input)
             throws CommentNotFoundException;
 }
